Fix: Issue 888
When projects are created inside a Maven Multi Module Project, they are added to the POM of maven multi module projects as modules. It was proposed to create profiles inside maven project's POM.
Hence, this commit changes the way that projects are added as modules. Now some projects are direclty added to module list of the root of POM, but some are added to the module lists of individual profiles.