Fix: Issue 946
When there is an array in the output whose elements are only assigned constant operator, then the output array is not mapped to any input array.
Therefore, there are no outputarrayloop declared for the array leaving nothing for constant operation to be assigned (Usually, each operation should be assigned to an arrayLoopBean for them to generate JS codes).

Fix: If there are any unassigned constant operation, declare outputarrayloop for the parent arrays of the outputs of the constant operations.