Fix: CompositeApp POM fails to open
When creating a synapse API from a swagger definition, if the API name contains a space, then the API artifact name will also contains a space. CompositeAPP POM does not support artifact names with space.
Fix: Replace space with "_" in the API name and replace space with _ in the swagger definition file name when creating a registry resource.