Fix issue with comment mediators getting deleted
As comment mediators does not have a graphical view, it is not shown
in the design view. Comment mediators are added to the output connector
of the previous mediator and it is attached to the previous mediator.
When the previous mediator is deleted, the comment mediator would get
deleted due to this reason.
Therefore, when deleting the mediators, the comment mediators that has
been attached to the output connector of the mediator is re attached
to the output connector of the mediator which is connected to the mediator
getting deleted.