Fix Inconsistency in Data Service creation wizard
The 'Generate Data Service from Datasource' page is shown regardless of the selection if we first
select the 'Generate Data Service from Datasource' radio button.
To fix this we need to reset the 'isGenerateDataService' flag when the other radio buttons are clicked.

Git issue: https://github.com/wso2/api-manager/issues/1597